requirements.txt
Flask==2.3.3
Markdown==3.7
MarkupSafe==2.1.5
openai==1.55.0
python-dotenv==1.0.1
gunicorn==20.1.0